refactoring unit tests and integration tests
