+++ tests/unit/test_tree.py
@@ -4,6 +4,28 @@ import pandas as pd
 from randomforests.Tree import DecisionTree
 
 
+def test_init():
+
+    tree = DecisionTree(3,2,1)
+
+    assert (tree.max_depth == 3 and
+            tree.min_size  == 2 and
+            tree.n_features == 1)
+
+def test_set_features():
+    tree = DecisionTree()
+    X    = pd.DataFrame({"x1":[0,1],"x2":[1,0]})
+    tree._set_features(X)
+
+    assert tree.n_features == 2
+
+def test_set_features_error():
+    tree = DecisionTree(n_features=5)
+    X    = pd.DataFrame({"x1":[0,1],"x2":[1,0]})
+    with pytest.raises(Exception):
+        tree._set_features(X)
+
+
 test_split_data =[
   ( np.array([[1, 4, 3, 0],
               [2, 3, 3, 0],
@@ -58,25 +80,4 @@ def test_private_predict(row, node, expected):
   result = tree._predict(row = row, node = node)
   assert expected == result
 
-
-def test_init():
-
-  tree = DecisionTree(3,2,1)
-
-  assert (tree.max_depth == 3 and
-          tree.min_size  == 2 and
-          tree.n_features == 1)
-
-def test_set_features():
-  tree = DecisionTree()
-  X    = pd.DataFrame({"x1":[0,1],"x2":[1,0]})
-  tree._set_features(X)
-
-  assert tree.n_features == 2
-
-def test_set_features_error():
-  tree = DecisionTree(n_features=5)
-  X    = pd.DataFrame({"x1":[0,1],"x2":[1,0]})
-  with pytest.raises(Exception):
-    tree._set_features(X)
